refactor(P3.9-4): delete the descriptor path — one builder, one IMD reader
PLAN.md 3.9 step 4. The second IMD→sequence builder in the Python binding
is gone: `resolve_algorithm_sequence`, `AlgorithmDescriptor`, the twelve
building-block pyclasses and the preset bodies (`algorithm_sequence.rs`
1473 → 101 lines). What survives one release are the deprecated names —
`AlgorithmSequence.nve/nvt/npt/minimize/from_parameters(topo, params)` now
return the `Plan` of the parameters through `gromos_run::build_plan`, and
`Simulation.from_sequence` takes that `Plan`. `Simulation.recipe`/`.plan`
are always present.

One reader: `gromos_run::read_imd` — the `md` binary, `Recipe.from_imd`,
`InputParameters` and the bundle loader all open parameter files through
it. `just lint` now runs the G6 drift gates after clippy (no
`AlgorithmSequence::new()` outside gromos-run, no `.push(Box::new(` in the
binding, no `read_imd_file(`/`parse_imd_str(` outside gromos-run/gromos-io,
no `process::exit` in gromos-run) — all clean.

Also: `Vec3`/`Frame`/`rmsd`/`rdf` are f64 (gromos_core::math::Vec3);
recipe/atom-count errors are `RecipeError`; `md_runners.py`, the
`test_basic.py`/`test_advanced_features.py` placeholders and the `glam`
dependency of the binding are deleted; `analysis` leaves the default
namespace; API reference, learning guide and the four CONTEXT files
describe the recipe as the only entry point. `test_front_end_parity.py`
loses path C and its twelve xfail rows; the deprecated names are checked
as an exact translation against the Recipe path instead.

Verified: Python 300 passed / 8 skipped / 3 xfailed, stubtest clean, G6
clean, gromos-run/pyo3-gromos tests, Rust reference suite 38 passed /
2 ignored, clippy without new findings.

+++ crates/pyo3-gromos/src/lib.rs
@@ -16,7 +16,7 @@
 //! gromos-* crates (pure Rust)
 //! ```
 
-use glam::Vec3A as Vec3;
+use gromos_core::math::Vec3;
 use numpy::{PyArray1, PyArray2, PyReadonlyArray2};
 use pyo3::prelude::*;
 
@@ -47,32 +47,32 @@ pub struct PyVec3 {
 #[pymethods]
 impl PyVec3 {
     #[new]
-    pub fn new(x: f32, y: f32, z: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64) -> Self {
         Self {
             inner: Vec3::new(x, y, z),
         }
     }
 
     #[getter]
-    pub fn x(&self) -> f32 {
+    pub fn x(&self) -> f64 {
         self.inner.x
     }
 
     #[getter]
-    pub fn y(&self) -> f32 {
+    pub fn y(&self) -> f64 {
         self.inner.y
     }
 
     #[getter]
-    pub fn z(&self) -> f32 {
+    pub fn z(&self) -> f64 {
         self.inner.z
     }
 
-    pub fn length(&self) -> f32 {
+    pub fn length(&self) -> f64 {
         self.inner.length()
     }
 
-    pub fn dot(&self, other: &PyVec3) -> f32 {
+    pub fn dot(&self, other: &PyVec3) -> f64 {
         self.inner.dot(other.inner)
     }
 
@@ -104,7 +104,7 @@ impl PyVec3 {
         }
     }
 
-    pub fn __mul__(&self, scalar: f32) -> PyVec3 {
+    pub fn __mul__(&self, scalar: f64) -> PyVec3 {
         PyVec3 {
             inner: self.inner * scalar,
         }
@@ -236,9 +236,9 @@ impl PyFrame {
     }
 
     /// Get positions as numpy array (N x 3)
-    pub fn positions_array<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyArray2<f32>>> {
+    pub fn positions_array<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyArray2<f64>>> {
         let _n_atoms = self.positions.len();
-        let data: Vec<f32> = self
+        let data: Vec<f64> = self
             .positions
             .iter()
             .flat_map(|v| [v.x, v.y, v.z])
@@ -263,8 +263,8 @@ impl PyFrame {
 /// Calculate RMSD between two sets of positions
 #[pyfunction]
 pub fn rmsd<'py>(
-    positions: PyReadonlyArray2<'py, f32>,
-    reference: PyReadonlyArray2<'py, f32>,
+    positions: PyReadonlyArray2<'py, f64>,
+    reference: PyReadonlyArray2<'py, f64>,
 ) -> PyResult<f64> {
     let pos = positions.as_array();
     let ref_arr = reference.as_array();
@@ -279,9 +279,9 @@ pub fn rmsd<'py>(
     let mut sum_sq = 0.0f64;
 
     for i in 0..n_atoms {
-        let dx = (pos[[i, 0]] - ref_arr[[i, 0]]) as f64;
-        let dy = (pos[[i, 1]] - ref_arr[[i, 1]]) as f64;
-        let dz = (pos[[i, 2]] - ref_arr[[i, 2]]) as f64;
+        let dx = pos[[i, 0]] - ref_arr[[i, 0]];
+        let dy = pos[[i, 1]] - ref_arr[[i, 1]];
+        let dz = pos[[i, 2]] - ref_arr[[i, 2]];
         sum_sq += dx * dx + dy * dy + dz * dz;
     }
 
@@ -292,7 +292,7 @@ pub fn rmsd<'py>(
 #[pyfunction]
 pub fn rdf<'py>(
     py: Python<'py>,
-    positions: PyReadonlyArray2<'py, f32>,
+    positions: PyReadonlyArray2<'py, f64>,
     group1: Vec<usize>,
     group2: Vec<usize>,
     n_bins: usize,
@@ -307,9 +307,9 @@ pub fn rdf<'py>(
     for &i in &group1 {
         for &j in &group2 {
             if i != j {
-                let dx = (pos[[i, 0]] - pos[[j, 0]]) as f64;
-                let dy = (pos[[i, 1]] - pos[[j, 1]]) as f64;
-                let dz = (pos[[i, 2]] - pos[[j, 2]]) as f64;
+                let dx = pos[[i, 0]] - pos[[j, 0]];
+                let dy = pos[[i, 1]] - pos[[j, 1]];
+                let dz = pos[[i, 2]] - pos[[j, 2]];
                 let r = (dx * dx + dy * dy + dz * dz).sqrt();
 
                 let bin = (r / dr) as usize;
